Move PasswordManager from feature to auth infra
