+++ server/application/src/main/kotlin/io/vitalir/kotlinhub/server/app/feature/user/domain/usecase/impl/RegisterUserUseCaseImpl.kt
@@ -7,7 +7,7 @@ import arrow.core.right
 import io.vitalir.kotlinhub.server.app.feature.user.domain.model.User
 import io.vitalir.kotlinhub.server.app.feature.user.domain.model.UserCredentials
 import io.vitalir.kotlinhub.server.app.feature.user.domain.model.UserError
-import io.vitalir.kotlinhub.server.app.feature.user.domain.password.PasswordManager
+import io.vitalir.kotlinhub.server.app.infrastructure.auth.PasswordManager
 import io.vitalir.kotlinhub.server.app.feature.user.domain.persistence.UserPersistence
 import io.vitalir.kotlinhub.server.app.feature.user.domain.usecase.RegisterUserUseCase
 import io.vitalir.kotlinhub.server.app.feature.user.domain.validation.UserValidationRule
